# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Production server",
        "title": "Carbon API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/generate-pdf": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generate a PDF report based on emission data and send it via email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Generate a PDF report for carbon emissions",
                "responses": {
                    "200": {
                        "description": "PDF generated and sent to fr081938@gmail.com",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error generating PDF or sending email",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/generate-pdf-summary": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generate a summary PDF report with emission data and AI-generated predictions, then send it via email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Generate a summary PDF report",
                "responses": {
                    "200": {
                        "description": "PDF generated and sent to fr081938@gmail.com",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error generating PDF or sending email",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}